backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
aiogram==3.4.1
gspread==6.0.0
google-auth==2.27.0
qrcode[pil]==7.4.2
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.27.0
pyzbar==0.1.9
pillow==10.2.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
sqladmin==0.16.0
itsdangerous==2.1.2
